tweaks

+++ experiments/trainers/ppo_rlhf_trainer.py
@@ -120,7 +120,8 @@ class PPORLHFTrainer(BaseTrainer):
 
         loss_ppo -= self.config.beta * entropy
 
-        return loss_ppo.half(), entropy.half()
+        # return loss_ppo.half(), entropy.half()
+        return loss_ppo, entropy.half()
     
     @profile
     def _backward(self, loss_value, loss_ppo):
